Fix communicator

Format code


Format code


Format code


Add touch events


Use better types


Use scrollTo instead of scrollIntoView


Lazyload emoji picker


don't post cursors if not changed


Fix duplicate message communicator


Extract page content
