+++ src/components/editor-page/editor-pane/tool-bar/emoji-picker/emoji-picker-button.tsx
@@ -38,3 +38,5 @@ export const EmojiPickerButton: React.FC = () => {
     </Fragment>
   )
 }
+
+export default EmojiPickerButton
